+++ src/middleware/defaultheaders.rs
@@ -150,7 +150,7 @@ where
             }
             Ok(res)
         }
-            .boxed_local()
+        .boxed_local()
     }
 }