+++ src/middleware/errhandlers.rs
@@ -140,7 +140,7 @@ where
                 Ok(res)
             }
         }
-            .boxed_local()
+        .boxed_local()
     }
 }