+++ src/types/form.rs
@@ -365,7 +365,7 @@ where
                         .map_err(|_| UrlencodedError::Parse)
                 }
             }
-                .boxed_local(),
+            .boxed_local(),
         );
         self.poll(cx)
     }